Added doc

See merge request lcom-25/pl/grupo_2leic04_6!45

+++ proj/src/model/score/score.h
@@ -7,6 +7,14 @@
 #include "sprite.h"
 #include "spriteLoader.h"
 
+/**
+ * @defgroup score score
+ * @brief Score management in the game
+ * 
+ * This module provides functions to create, manage, and draw the score in the game.
+ * It handles the current score, increments it, and draws the score on the screen.
+ */
+
 struct score_imp;
 typedef struct score_imp Score;
 
@@ -17,14 +25,30 @@ typedef struct score_imp Score;
  * @param y The y-coordinate of the score
  * @return Pointer to the created Score object, or NULL if the creation failed
  */
-
 Score *create_score(SpriteLoader *loader);
 
+/**
+ * @brief Destroys a Score object, freeing all allocated memory
+ * 
+ * @param score Pointer to the Score object to be destroyed
+ */
 void destroy_score(Score *score);
 
+/**
+ * @brief Increments the score by 1
+ * 
+ * The score is incremented every 30 frames.
+ * 
+ * @param score Pointer to the Score object
+ */
 void increment_score(Score *score);
 
-uint8_t get_score_value(Score *score);
+/**
+ * @brief Gets the current score value
+ * 
+ * @return The current score value, or 0 if the score is NULL
+ */
+uint16_t get_score_value(Score *score);
 
 void draw_score(Score *score, int x, int y);
